Move FunctionLengthSniff to Functions namespace
