+++ tests/Sniffs/Functions/data/functionLengthErrors.php
