+++ tests/Sniffs/Functions/data/functionLengthNoErrors.php
